requirements.txt
aiogram==2.25.1
aiosqlite==0.19.0
aiohttp==3.8.5
python-dotenv==1.0.0
nest-asyncio==1.5.7
pydantic==2.1.1
openai==0.27.8
matplotlib==3.7.2
aioschedule==0.5.2
Pillow==10.0.0
loguru==0.7.0
httpx==0.24.1
python-dateutil==2.8.2